Add ability to toggle sorting bag info keys
In certain cases, a user may want the tags to be written to the
bag-info.txt in a specific order.

In my own uses for BagIt, I am writing tags to the bag info according to
a certain archival accession standard (CAAIS) which has a specific
ordering to the metadata fields. Without being able to turn off sorting
the keys, the different fields of data in different sections are all
mixed together and it is confusing to read.

Here's an example:
I want all of the user's contact information to appear in order in the
bag-info.txt file. I am writing and "addressLine1" tag, an "email" tag,
and a "phoneNumber" tag (among others). Now, I also need to write an
"archivalUnit" tag, and a "generalNote" tag. Using the default
behaviour, I get:

addressLine1: ...
archivalUnit: ...
email: ...
generalNote: ...
phoneNumber: ...

The contact information is mixed in with all of the other information!
But if I now pass sort_keys=False when I create the bag, all my problems
are solved.

The reason I used kwargs is so that the bag creation behaviour is not
deprecated for anyone else, the kwargs in my opinion are a safe way to
add more arguments without polluting the call signature too much.